Spring cleaning
* Added final modifiers
* Remove redundant casts
* Fixed deprecated code
* Fixed potential bugs

+++ src/main/java/com/aether/blocks/natural/AetherLeafPileBlock.java
@@ -21,7 +21,7 @@ public class AetherLeafPileBlock extends FallingBlock {
     @Override
     public void onEntityCollision(BlockState state, World world, BlockPos pos, Entity entity) {
         if (canFallThrough(world.getBlockState(pos.down()))) {
-            FallingBlockEntity fallingBlockEntity = new FallingBlockEntity(world, (double)pos.getX() + 0.5D, (double)pos.getY(), (double)pos.getZ() + 0.5D, world.getBlockState(pos));
+            FallingBlockEntity fallingBlockEntity = new FallingBlockEntity(world, (double)pos.getX() + 0.5D, pos.getY(), (double)pos.getZ() + 0.5D, world.getBlockState(pos));
             this.configureFallingBlockEntity(fallingBlockEntity);
             world.spawnEntity(fallingBlockEntity);
         }
